feat: Lazy-load animated icons as a separate bundle chunk
Animated SVG data was bundled unconditionally into the single JS file
even when animated_icon is false. On low-power devices (e.g. NSPanel Pro
with ARM Cortex-A35) this caused the main thread to block long enough to
drop the HA WebSocket keepalive, making the dashboard freeze on load.

Changes:
- Split src/images.ts: animated imports/export moved to animatedImages.ts,
  static icons remain in images.ts
- Dynamic import('./animatedImages') is triggered in setConfig only when
  animated_icon != false; static icon is shown as fallback until loaded
- rollup.config.js: manualChunks + chunkFileNames produce a fixed-name
  animated-icons.js chunk alongside the main bundle
- hacs.json: add zip_release: true so HACS deploys both files from the
  release zip instead of only the single main JS
- release workflow: zip both files into clock-weather-card.zip for upload

Result: main bundle 204 KB (was 302 KB, -33%), animated icons load
asynchronously in the background only when the feature is enabled.
